+++ CHANGELOG.md
@@ -1,18 +1,52 @@
 # Change History for Saropa Dart Utils
 
+## 0.0.11 - Rome (Latest)
+
+* 🗑️ Removed deprecated functions in ```StringNullableExtensions```
+
+* 🐛 Fixed ```StringExtensions.removeStart``` to return the input string when the search param is empty
+ ```'Hello, World!'.removeStart(''); // 'Hello, World!'```
+
+* ⚙️ Added constant ```DateConstants.unixEpochDate```
+ ```DateConstants.unixEpochDate; // January 1st, 1970```
+
+* ✨ Added function ```DateConstantExtensions.isUnixEpochDate```
+ ```DateTime.utc(1970).isUnixEpochDate; // true```
+
+* ✨ Added function ```DateConstantExtensions.isUnixEpochDateTime```
+ ```DateTime.utc(1970, 1, 1, 0, 0, 1).isUnixEpochDateTime; // false```
+
+* ✨ Added function ```IntStringExtensions.ordinal```
+ ```101.ordinal(); // 101st```
+
+* ✨ Added function ```StringUtils.getNthLatinLetterLower```
+ ```String? StringUtils.getNthLatinLetterLower(3) // "c"```
+
+* ✨ Added function ```StringUtils.getNthLatinLetterUpper```
+ ```String? StringUtils.getNthLatinLetterUpper(4) // "D"```
+
+* ✨ Added function ```IntUtils.findGreatestCommonDenominator```
+ ```String? IntUtils.findGreatestCommonDenominator(15, 45) // 15```
+
+* ✨ Added function ```IntExtensions.countDigits```
+ ```(-12345).countDigits() // 5```
+
 ## 0.0.10 - Paris
 
-* Change start to be nullable ```String? removeStart(String? start)```
+* Change start to be nullable
+```String? removeStart(String? start)```
 
 ## 0.0.9 - Geneva
 
 * Review of meta data
-* Add trimFirst param to ```StringExtensions.removeStart({bool trimFirst = false,}){...}```
+* Add trimFirst param to
+```StringExtensions.removeStart({bool trimFirst = false,}){...}```
 
 ## 0.0.8 - Nepal
 
 * Review of meta data
-* Add trimFirst param to ```StringExtensions.nullIfEmpty({bool trimFirst = true,}){...}```
+* Add trimFirst param to
+```StringExtensions.nullIfEmpty({bool trimFirst = true,}){...}```
 
 ## 0.0.7 - Jamaica